Design and analysis a carframe for FSAE.
Formula Student China is a hole nation competion for university students in China.  We will design and manufactory our own racing car and take part in the competion. The carframe is a very important part for the racing car. It bear the weight of almost all the other part of the racing car. And proctect them, include the driver. So the design and analysis of the carfame is very important. This project will desing a carframe for SUSTech Nanshan Racing car Team.